feat(pathtools): Refactor method in `pathtools`
* Implement function to join filenames with channel suffixes.
* Suitable for Bioformats/Jython in show_list mode.
* Includes detailed docstring with parameters and return values.
